Server side: Refactor

+++ server-side/controller/UserController.js
@@ -1,21 +1,6 @@
 const UserRepository = require('../repository/UserRepository');
-const { validationResult, body } = require('express-validator/check');
-
-exports.validate = () => {
-  return [
-    body('id', 'id doesnt exist').exists(),
-    body('firstname', 'firstname doesnt exist').exists(),
-    body('lastname', 'lastname doesnt exist').exists(),
-    body('email', 'email doesnt exist').exists().isEmail(),
-    body('birthDate', 'birthDate doesnt exist').exists(),
-    body('address', 'address doesnt exist').exists(),
-    body('address.id', 'address doesnt exist').exists(),
-    body('address.street', 'street doesnt exist').exists(),
-    body('address.city', 'city doesnt exist').exists(),
-    body('address.country', 'country doesnt exist').exists(),
-    body('address.postalcode', 'postalcode doesnt exist').exists(),
-  ]
-}
+const { validationResult } = require('express-validator/check');
+const errorMessages = require('../utils/errorMessages');
 
 exports.getAll = (req, res) => {
   console.log('UserController ---- getAll');
@@ -24,10 +9,14 @@ exports.getAll = (req, res) => {
 }
 
 exports.getById = (req, res) => {
+  console.log('UserController ---- getById');
+  if (!req.params.id) {
+    return res.status(400).json(errorMessages.badRequest());
+  }
   const id = parseInt(req.params.id, 10);
   UserRepository.getById(id)
     .then(data => {
-      (data) ? res.status(200).json(data) : res.status(404).json({error: "User not found"});
+      (data) ? res.status(200).json(data) : res.status(404).json(errorMessages.notFound('userID'));
     });
 }
 
@@ -49,22 +38,25 @@ exports.updateById = (req, res) => {
   
   const validation = validationResult(req);
   if (!validation.isEmpty()) {
-    return res.status(405).json({ errors: validation.array() });
+    return res.status(400).json({ errors: validation.array() });
   } 
 
   UserRepository.updateById(id, req.body)
     .then(data => {
-      (data) ? res.status(200).send(data) : res.status(404).json({error: "User not found"});
-    })
+      (data) ? res.status(200).send(data) : res.status(404).json(errorMessages.notFound('userID'));
+    });
 }
 
 exports.deleteById = (req, res) => {
   console.log('UserController ---- delete');
+  if (!req.params.id) {
+    return res.status(400).json(errorMessages.badRequest());
+  }
   const id = parseInt(req.params.id, 10);
   
   UserRepository.deleteById(id)
     .then(data => {
-      console.log('finish' + data);
-      (data !== -1) ? res.status(200).json({success: "true"}) : res.status(404).json({error: "User not found"});
+      (data !== -1) ? res.status(200).json({success: 'true'}) : res.status(404).json(errorMessages.notFound('userID'));
     });
 }
+
